# opts_backend_code/saasbasedcustomdashboard/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Saas API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/google-spreadsheet-list": {
            "get": {
                "tags": [
                    "GoogleSheets"
                ],
                "description": "Fetch the list of all sheets present on user google account",
                "operationId": "bcf36e4f9e35e58663fe71ccdf290b33",
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/fetch-spreadsheet-sheets-list/{id}": {
            "get": {
                "tags": [
                    "GoogleSheets"
                ],
                "description": "Fetch the sheets present within the spreadsheet",
                "operationId": "c6fe46486948f6efb879046245381b6f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Spread sheet Id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/fetch-sheet-headers/{id}": {
            "get": {
                "tags": [
                    "GoogleSheets"
                ],
                "description": "",
                "operationId": "92f4621bdd8f8b610237cf87f02d4464",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Spreadsheet Id",
                        "required": true
                    },
                    {
                        "name": "sheet_name",
                        "in": "query",
                        "description": "Enter the selected sheet name",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}